Move misc functions to own compile unit
- clean up definitions of function macros
